requirements.txt
# Dataset
datasets>=2.19
huggingface-hub>=0.23
pandas>=2.0
pyarrow>=15.0

# Training
torch>=2.1
transformers>=4.45
peft>=0.12
trl>=0.9
accelerate>=0.33
unsloth

# Evaluation
requests>=2.31
numpy>=1.24
scikit-learn>=1.3
tabulate>=0.9
seqeval>=1.2.2
evaluate>=0.4

# Utilities
tqdm>=4.66
pyyaml>=6.0
